refactor: make module private
